+++ styles/all/template/hangman_input.js
@@ -21,24 +21,39 @@ $("#hangman_quote_input").submit(function() {
 	// delete multiple spaces
 	elementValue = elementValue.replace(/ {2,}/g, ' ');
 
-	// check whether only authorized letters are used
+	// check whether only authorized letters are used and count the authorized letters
 	var strLenght = elementValue.length;
 	var onlyAuthLetters = true;
 	var checkLetter;
+	var countLetters = 0;
 	var unauthLetters = new Array();
 	for (var i = 0; i < strLenght; i++) {
+		// check one character after another
 		checkLetter = elementValue[i];
+		// check for unauthorized characters
 		if (!motHangman.jsLcLetters.includes(checkLetter.toLowerCase()) && !motHangman.jsAllowedPunctMarks.includes(checkLetter) && checkLetter != ' ') {
 			unauthLetters.push(checkLetter);
 			onlyAuthLetters = false;
 		}
+		//count authorized letters
+		if (motHangman.jsLcLetters.includes(checkLetter.toLowerCase())) {
+			countLetters++;
+		}
 	}
+
 	$("#hangman_quote_text").val(elementValue);
 	if (!onlyAuthLetters) {
-		alert (motHangman.jsUnauthLetters + unauthLetters);
+		phpbb.alert (motHangman.jsErrorTitle, motHangman.jsUnauthLetters + unauthLetters);
 		$("#hangman_quote_text").focus();
 	}
+
 	if (elementValue == '' || elementValue == ' ') {
+		$("#hangman_quote_text").focus();
+		onlyAuthLetters = false;
+	}
+
+	if (countLetters < motHangman.jsMinTermLen) {
+		phpbb.alert (motHangman.jsErrorTitle, motHangman.jsTermTooShort + motHangman.jsMinTermLen);
 		$("#hangman_quote_text").focus();
 		onlyAuthLetters = false;
 	}
@@ -51,7 +66,7 @@ $("#hangman_quote_input").submit(function() {
 		elementValue = elementValue.replace(/ {2,}/g, ' ');
 
 		if (motHangman.jsEnforceCategory && elementValue == '') {
-			alert('Kategorie muss ausgefüllt sein');
+			phpbb.alert(motHangman.jsErrorTitle, motHangman.jsCategoryMissing);
 			$("#hangman_quote_category").focus();
 			onlyAuthLetters = false;
 		}